design(workstation): split THEME_PRESET_COLORS data out of theme.ts
theme.ts held two unrelated things at 2,579 lines: the LogInkTheme
resolver logic and the ~128-entry color preset catalog. The data
dwarfed the logic, so every theme addition churned a file that also
owns behavior, and reviewing a one-entry theme PR meant scrolling a
resolver module.

Pure data move: THEME_PRESET_COLORS now lives in themePresets.ts,
re-exported from theme.ts so every existing import site is unchanged.
Updated AGENTS.md's theme rule and the workstation README's directory
map to point at the new file.

Fixes #1640.
